requirements.txt
streamlit
python-dotenv
requests
groq
pandas